feat: Introduce simple example project with OpenAPI integration
- Added a new example project demonstrating the usage of the API server with TypeScript.
- Implemented OpenAPI generation for user management routes, including CRUD operations.
- Updated package dependencies in pnpm-lock.yaml to support the new example.
- Created necessary configuration files (package.json, tsconfig.json) for the example project.
- Enhanced the compiler with a new OpenAPI generator for improved API documentation.
